requirements.txt
# Core AI/ML frameworks
langchain==0.1.0
langchain-anthropic==0.1.0
langgraph==0.0.20

# MCP (Model Context Protocol)
mcp==0.9.0

# Data validation and processing
pydantic==2.5.0
pandas==2.1.4

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.26.0

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# CLI enhancements
rich==13.7.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
